% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acm_operations.R
\name{export_certificate}
\alias{export_certificate}
\title{Exports a private certificate issued by a private certificate authority (CA) for use anywhere}
\usage{
export_certificate(CertificateArn, Passphrase)
}
\arguments{
\item{CertificateArn}{An Amazon Resource Name (ARN) of the issued certificate. This must be of the form:

\code{arn:aws:acm:region:account:certificate/12345678-1234-1234-1234-123456789012}}

\item{Passphrase}{Passphrase to associate with the encrypted exported private key. If you want to later decrypt the private key, you must have the passphrase. You can use the following OpenSSL command to decrypt a private key:

\code{openssl rsa -in encrypted_key.pem -out decrypted_key.pem}}
}
\description{
Exports a private certificate issued by a private certificate authority (CA) for use anywhere. You can export the certificate, the certificate chain, and the encrypted private key associated with the public key embedded in the certificate. You must store the private key securely. The private key is a 2048 bit RSA key. You must provide a passphrase for the private key when exporting it. You can use the following OpenSSL command to decrypt it later. Provide the passphrase when prompted.
}
\details{
\code{openssl rsa -in encrypted_key.pem -out decrypted_key.pem}
}
\examples{

}
