% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acm_operations.R
\name{list_tags_for_certificate}
\alias{list_tags_for_certificate}
\title{Lists the tags that have been applied to the ACM certificate}
\usage{
list_tags_for_certificate(
  CertificateArn = "string"
)
}
\arguments{
\item{CertificateArn}{[required] String that contains the ARN of the ACM certificate for which you want to list the tags. This must have the following form:

\code{arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012}

For more information about ARNs, see \href{http://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resource Names (ARNs) and AWS Service Namespaces}.}
}
\description{
Lists the tags that have been applied to the ACM certificate. Use the certificate's Amazon Resource Name (ARN) to specify the certificate. To add a tag to an ACM certificate, use the AddTagsToCertificate action. To delete a tag, use the RemoveTagsFromCertificate action.
}
\examples{

}
