% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acm_package.R
\docType{package}
\name{paws.acm-package}
\alias{paws.acm}
\alias{paws.acm-package}
\title{paws.acm: AWS Certificate Manager}
\description{
Welcome to the AWS Certificate Manager (ACM) API
    documentation.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
